Add package management system and Projectile
This will automatically check if all the required packages (currently,
just Projectile) are installed. If not, any missing packages will be
automatically downloaded. Having a proper setup like this instead of
using M-x package-install allows radon-emacs to be deployed on an
arbitrary machine without additional configuration.

The package cache and the Projectile project cache have been added to
.gitignore. Additionally, Emacs seems to be using GnuPG now, and those
files have been placed in this directory since it has to be set as $HOME
in order to support multiple swappable Emacs configurations. So, .gnupg
has gone into the .gitignore as well.