Configure automatic backups for Emacs
Put backup files in $HOME/.emacs-backups. If you're running this setup
as a swappable configuration (i.e. resetting $HOME when starting Emacs)
then the .emacs-backups directory will be placed in this project. Hence
the addition to .gitignore.

The ido.last~ entry has also been removed from .gitignore, because it
was an automatic backup and these have been moved to .emacs-backups.