Fix for hiding CIDER mode lighter
I hadn't realized that there was a CIDER-specific variable for the CIDER
mode lighter. Generally speaking, if a package provides a customizable
variable for its mode lighter, any changes you make directly to
minor-mode-alist will be overridden (although not always) when the mode
loads.

So, this change makes CIDER's mode lighter hide properly.