Remove 'subtle option for radon-emacs-tweak-colors
This option didn't really make a whole lot of sense, because all color
tweaks are either going to be hardcoded or based on a preexisting face.
In the former case, we want them *off* when not setting a color theme;
in the latter case, we want them *on* even when not setting a color
theme. There is no in-between.