Don't make assumptions about the $PATH in setup.sh
Previously, the setup script could get messed up if you had an old
version of one of the executables installed to /usr/local/bin earlier on
your $PATH. Now, it will check for this and create an alias for itself,
if necessary. Furthermore, all aliases will be saved to the
.zshrc.aliases file in the scripts directory. This file can then be
sourced from the main .zshrc.

Also, the code for installing Emacs has been made much more robust with
regard to dealing with broken symlinks and non-EmacsForMacOSX versions
of Emacs.