Bugfix for alembic/distill outside of projects
Previously, you couldn't pull from *any* external repositories -- even
the default ones, Central and Clojars -- when outside a project. This
was because the hack for allowing distill to work outside of projects
defaulted the repository list to the empty vector. Replacing the default
repository list with a vector containing the Central and Clojars
repositories fixes the issue.