Account for broken Homebrew symlinks in setup
Also, add postconditions to all the installer scripts -- if the command
is not accessible even after we installed it, just abort. Something is
going on that is too weird to try and diagnose automatically, at least
for now. Since it looks like /usr/local/bin should be on the $PATH by
default, and earlier than /usr/bin and /bin, there should be no need for
creating aliases. After all, the $PATH will be reset to the default as
soon as .zshrc is replaced and a new shell session is started.