Use native tmux instead of powerline
Many hours were wasted in the environment/configuration hell that is
Powerline. Let me be straight here: everything is overcomplicated and
the documentation is worthless. The whole thing can be done in just a
few lines of .tmux.conf without introducing multiple external
dependencies (Python, pip, powerline, whatever you need for the
default(!) segments to work), counterintuitive and hard-to-find
configuration files, and extra symlinks.