Refactor ensure-symlinked.sh to fix error
Previously, the script tried to make $link an absolute path before
ensuring that intermediate directories in its path existed. This caused
an error when the intermediate directories did not exist.

This commit refactors the whole script to fix this, and also makes other
general improvements, such as adding more comments and improving the
clarity of messages output to the user.

Resolves #95.