Allow usage of the C-c M-J shortcut everywhere
The way this code was written previously caused some odd behavior. In
particular, if you pressed C-c M-J outside a Clojure file, then the
shortcut would not be available, and Emacs would automatically translate
it to C-c M-j. This caused just a regular Clojure REPL to be opened,
instead of a ClojureScript one! Adding the C-c M-J shortcut explicitly
avoids this problem.