Replace ../local with ../../dotfiles-local
If the local config files are in the main repository, then they have to
be in .gitignore. But unless you use native indexing (which is slow),
then Projectile has no way of forcing certain ignored files to be shown.
So you can't reach your local config files with Projectile, which sucks.
The only solution (that I see so far) is to put the local config files
in a separate directory (which could optionally be made a Git repository
in its own right).

The other change in this commit is to replace ${real#../} with just
$real in ensure-symlinked.sh. The former version was a bit of a hack,
and it would have needed to be even more complicated in order to support
paths with more than one '..' in them. Better to just report the actual
true path to the user.