Tweak tweakable parameter system in .zshrc
Now the parameters are only required to be != false, rather than ==
true, in order for the settings they control to take effect. This
eliminates the need for the section at the top, which reduces code
duplication and increases maintainability. Hopefully, coming changes to
the setup script will make it just as easy to find out what things can
be changed as it was with the section at the top.