Make setup script version parsing more permissive
This commit modifies ensure-installed.sh to make the space after the
$version_command_name but before the version number *optional*. This
allows proper parsing of version strings that go like 'Racket v6.6'.

Note that we still check the well-formed-ness of the resulting version
string (in the above example, '6.6'), so this additional permissiveness
does not introduce any fragility with regard to malformed version
strings.