Revert "Allow shell functions and aliases in M-!"
It turns out that this causes a noticeable slowdown when killing text
(with e.g. C-k or C-w), since a shell command is run to send the text to
the system clipboard. This is a *major* UX problem, so I'm reverting the
change until we can figure out how to tell Emacs to not use the -i flag
when copying to the system clipboard, or more generally in any situation
*except* M-!.