Remove Magit
I'm not using Magit because issues #51 and #52 are still unresolved, so
it makes sense to remove it. If you want to keep Magit, simply add the
following line to your init.before.local.el:

(radon-add-package 'magit)

Or, you can set up init.before.local.el interactively by deleting any
existing version of the file and re-running the setup script as follows:

$ scripts/setup.sh only emacs

Please note that if you already had Magit installed by package.el (for
instance, if you had this configuration set up already, and you are
upgrading to a version that includes this commit), you will not notice
the change immediately, since removed packages are not pruned eagerly.
However, if you delete ~/.emacs.d/elpa, you will have to add the above
line for Magit to be re-downloaded automatically.