Improve init.el parameter parsing
This commit is primarily designed to address the problem of random lines
that look like "(setq radon-..." being detected as configuration
parameters by the interactive init.before.local.el creator in the setup
script. This is accomplished by making all the configuration parameters
start with the prefix "radon-customize-".

The relevant change to the create-init-before-local-el.sh parser has
also been made. This changes allows us to get rid of the radon-packages
special case in the parser (which was never actually necessary in the
first place, but was presumably introduced to be extra safe -- a
precaution that is no longer necessary when the configuration parameters
all have a dedicated prefix).

Resolves #130.