Allow radon-packages to be reloaded
This commit uses the defvar-nil-setq pattern described in the ErgoEmacs
page linked by the patch to make it so that when changes are made to
radon-packages and init.el is reloaded, those changes will actually show
up (e.g., any new packages will be loaded).

The necessary changes have also been made to the parser in
create-init-before-local-el.sh.