Rename dotfiles-local to radian-local
This is for consistency with the renamining of the project as a whole
from raxod502/dotfiles to Radian.

Code has been included for backwards compatiblity (i.e. if a
radian-local folder doesn't exist but a dotfiles-local folder does, then
the setup script will offer to rename the latter to the former. This can
be removed eventually to decrease code complexity.