Fix misplacement of Company-REPL conflict fix
Company overrides the M-p and M-n bindings that are supposed to be
available in the CIDER and Geiser REPLs. Originally, Radian only
included CIDER and therefore the logic was located in the CIDER section
(but was only run if Company was also loaded). When Geiser was added,
this code was extended to cover the Geiser REPL as well, but this was
done improperly because all the code was left in the CIDER section (so
that the Geiser REPL would not get the fix if CIDER was not included).

This commit moves the logic into the Company section, so that the fix is
applied separately to the CIDER and Geiser REPLs depending on which
one(s) is/are loaded.