Improve version string validation
Previously, the logic for ensuring version strings are well-formed was
duplicated between the 'is_installed_correctly' and 'install' functions.
It turned out that this duplication was done imperfectly, so that
'is_installed_correctly' knew how to deal with versions that had
modifiers like '-alpha' and '_1' postpended, but 'install' didn't.

This commit moves the logic for checking version string well-formedness
directly into the 'version_as_recent' function, so that both
'is_installed_correctly' and 'install' can take advantage of it.

This resolves a problem encountered when installing Python 3 from
Homebrew using ensure-installed.sh, because Homebrew postpends one of
the aforementioned qualifiers to the version it reports for python3, and
so 'install' choked when trying to recreate broken symlinks.