Add emacsc and emacscw for easy use of emacsclient
This commit also modifies the emacsw script to determine whether it
should dispatch to EmacsForMacOSX or $(which emacs) at runtime, instead
of setup time. This is simpler, and the same pattern is used for the new
emacsc and emacscw scripts.

In order to make this possible, we need a consistent way to get at the
Radian directory. This is done by creating a ~/.radian (and
~/.radian-local, for consistency) symlink.