Fix bug in ensure-not-version-controlled.sh
This commit modifies the logic that computes the name of the directory
to be (potentially) moved so that it resolves to the actual name of the
directory, not the name of the directory it points to (if it is a
symlink).

This makes it so we have full control over the names of things that are
put into the originals folder, which means we can be 100% confident that
nothing is going to get overridden.