Don't set "allow-rename" in .tmux.conf
It turns out that I had completely misunderstood the purpose of this
command, and it was not at all doing what its comment implied. The
functionality that I wanted to achieve by including the command (that is
to say, what is in the comment) is in fact what happens by default.