Make setup.sh handle profiles.clj more robustly
The previous setup had two problems: (1) if new dependencies had been
added to profiles.clj since the last time Leiningen had been run (which
would happen, for instance, if setup.sh was run twice in a row on a new
system), then checking the version of Leiningen would take a long time;
and (2) if profiles.clj was corrupt, checking the version of Leiningen
would fail entirely and the setup script would halt.

This commit works around both of these problems by moving profiles.clj
to the "originals" folder before checking the version of Leiningen, and
then moving it back. The symlinking step is left as is (so if the user
has a pre-existing profiles.clj, it will be moved to "originals", then
back again, then back to "originals"), for simplicity.