Re-enable CIDER's dependency injection
The problem with turning off dependency injection is that MELPA is
completely unversioned, whereas profiles.clj (even if you request a
SNAPSHOT build) is confined to a particular version (e.g.
0.15.0-SNAPSHOT). This means that if you check out an old version of
Radian, then you might get a version of the package.el CIDER that is too
new for the version of cider-nrepl specified in profiles.clj.

Turning off dependency injection fixes this. There's no need to remove
the entries in profiles.clj, since the injected versions of the
dependencies override the versions specified in profiles.clj. Plus,
leaving them in will allow other editors (which do not support
dependency injection) to use them.