Make 'git reword' alias more robust
Previously, if the commit failed (for instance, if you aborted using C-c
C-k in emacsclient, or if you submitted an empty message), then the
stash created by 'git reword' would not be popped. This commit fixes
that: the stash is always popped, so you will never enter a situation in
which it looks like your changes have been lost!