Fix extreme edge case with ivy-minibuffer-map
The bindings in this should be established when Ivy is loaded, not
when Counsel is loaded. This just bit me when I launched Emacs and
immediately invoked find-function on a C function, which prompted me
to find the Emacs C source directory using
read-directory-name (bypassing loading Counsel!), and then the
keybindings were not properly established (TAB was not working as
expected).