Update reattach-to-user-namespace hack
I decided to make it so that tmux can still start on macOS even if you
don't have reattach-to-user-namespace installed. While doing this, it
became clear that I was supposed to be using 'set' instead of 'tmux
set'. I have no idea why it was working before(?), although I did
update tmux from 2.2 to 2.4 recently. Luckily, tmux appears to
interpret environment variables such as $SHELL within the if-shell
construct.