Update all packages
Don't use fork of clj-refactor.el because that pull request is never
going to get merged and I don't really care about it anymore.

Don't use fork of ivy because it was a real pain to resolve the merge
conflicts, and maybe we can find a better solution that doesn't
involve a fork (or maybe the bug was accidentally fixed by someone).