Remove webpaste.el
I have a command-line script that does the same, except it uses
authentication so that I can later delete the pastes. Having the
option to create pastes that I can't delete is dangerous and is being
removed. If you depend on this functionality, either add it to your
local configuration or submit a request for the command-line
script (minus username and password) to be added to Radian.