Breaking change: migrate to znap
zinit was continually causing weird issues with plugins that marlon
richert told me were definitely the fault of zinit rather than the
plugins. decided to migrate to znap, and indeed it solves a lot of
weird behavior, in addition to being simpler and faster.

you need to change your radian_zinit_hook to radian_znap_hook if you
were using the former, and re-write code to work using znap syntax
instead of zinit. on the plus side, the new way is a lot simpler.

also removed fast-syntax-highlighting because the linked bug was
solved in zsh-autocomplete.