Fix mode line conflict with eldoc
If you would get eldoc triggered while in the minibuffer, for example
in eval-expression, the mode line in the buffer would get replaced
with a single percent-sign.

This was because it works by prepending some content to
mode-line-format, which breaks the :eval construct used by Radian.
But, since mode-line-format is evaluated recursively, we can just add
another layer of list and it works perfectly.