Using non-obsolete names (as of Emacs 31.0.50) (#556)
Note: For some reason, the PR is following the branch rather than the
commit.

I had wanted to propose both commits as PRs, so I will leave it as is.

The first commit makes some small changes to not use obsolete names in
Emacs 31.0.50

The second commit removes killing the compilation buffer of radian and
instead just resets its content. I probably recompile radian more than
average due when changing my config, and I typically have this buffer
open to see the warning messages and resolve them. Killing the buffer
and creating a new one means that I have to switch to the buffer after
every compilation, even if I had it opened and displayed before starting
the compilation.